requirements.txt
asgiref==3.8.1
Django==4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
setuptools==74.0.0
sqlparse==0.5.1
wheel==0.44.0
psycopg2-binary>=2.9
python-decouple
coverage==7.6.4
pytest==8.3.3
